requirements.txt
biopython==1.86
pandas==2.2.3
numpy==2.1.3
tqdm==4.67.1
sentence-transformers==3.4.1
torch==2.6.0
scikit-learn==1.6.1
umap-learn==0.5.7
hdbscan==0.8.40
plotly==6.0.0
keybert==0.8.5
